review fix

+++ internal/manager/orderManager.go
@@ -39,7 +39,7 @@ func (m *OrderManager) AddOrder(ctx context.Context, userID int64, orderNumber s
 	return nil
 }
 
-func (m *OrderManager) AccrualOrder(ctx context.Context, number string, status string, accrual float64) error {
+func (m *OrderManager) AccrualOrder(ctx context.Context, number string, status string, accrual uint) error {
 	if err := m.storage.UpdateOrder(ctx, number, status, accrual); err != nil {
 		return err
 	}
@@ -51,7 +51,7 @@ func (m *OrderManager) AccrualOrder(ctx context.Context, number string, status s
 	return err
 }
 
-func (m *OrderManager) WithdrawOrder(ctx context.Context, userID int, number string, withdraw float64) error {
+func (m *OrderManager) WithdrawOrder(ctx context.Context, userID int, number string, withdraw uint) error {
 	balance, err := m.storage.GetBalance(ctx, int64(userID))
 	if err != nil {
 		return err
